feat: added JWT authentication and authorization

+++ server/resolvers/VideoResolvers.js
@@ -5,38 +5,68 @@ const VideoResolvers = {
     hello: () => {
       return "hello world";
     },
-    getAllVideos: async () => {
-      try {
-        return await Video.find();
-      } catch (e) {
-        console.error({ error: e });
+    getAllVideos: async (parent, args, context, info) => {
+      const { isAuth, userId, email, name } = context;
+
+      // if(isAuth){
+
+      // }else{
+      //   throw new AuthenticationError("unauthorized access");
+      // }
+
+      if (isAuth) {
+        try {
+          return await Video.find();
+        } catch (e) {
+          console.error({ error: e });
+        }
+      } else {
+        throw new AuthenticationError("unauthorized access");
       }
     },
   },
   Mutation: {
     addVideo: async (parent, args, context, info) => {
-      try {
-        const { title, views, likes, link, category } = args.video;
-        const newVideo = new Video({ title, views, likes, link, category });
-        await newVideo.save();
-      } catch (e) {
-        console.error({ error: e });
+      const { isAuth, userId, email, name } = context;
+
+      if (isAuth) {
+        try {
+          const { title, views, likes, link, category } = args.video;
+          const newVideo = new Video({ title, views, likes, link, category });
+          await newVideo.save();
+        } catch (e) {
+          console.error({ error: e });
+        }
+      } else {
+        throw new AuthenticationError("unauthorized access");
       }
     },
     getVideo: async (parent, args, context, info) => {
-      try {
-        const { id } = args;
-        return await Video.findById(id);
-      } catch (e) {
-        console.error({ error: e });
+      const { isAuth, userId, email, name } = context;
+
+      if (isAuth) {
+        try {
+          const { id } = args;
+          return await Video.findById(id);
+        } catch (e) {
+          console.error({ error: e });
+        }
+      } else {
+        throw new AuthenticationError("unauthorized access");
       }
     },
     deleteVideo: async (parent, args, context, info) => {
-      try {
-        const { id } = args;
-        return await Video.findByIdAndDelete(id);
-      } catch (e) {
-        console.error({ error: e });
+      const { isAuth, userId, email, name } = context;
+
+      if (isAuth) {
+        try {
+          const { id } = args;
+          return await Video.findByIdAndDelete(id);
+        } catch (e) {
+          console.error({ error: e });
+        }
+      } else {
+        throw new AuthenticationError("unauthorized access");
       }
     },
   },
